APPEALS-43193: bigint

+++ db/migrate/20240507145931_add_columns_to_transcriptions.rb
@@ -1,9 +1,9 @@
 class AddColumnsToTranscriptions < Caseflow::Migration
   def change
     add_column :transcriptions, :transcription_contractor_id, :bigint
-    add_column :transcriptions, :created_by_id, :integer
+    add_column :transcriptions, :created_by_id, :bigint
     add_column :transcriptions, :transcription_status, :string, comment: "Possible values: 'unassigned', 'in_transcription', 'completed', 'completed_overdue'"
-    add_column :transcriptions, :updated_by_id, :integer
+    add_column :transcriptions, :updated_by_id, :bigint
     add_column :transcriptions, :deleted_at, :datetime, comment: "acts_as_paranoid in the model"
   end
 end
